Merge translations of Bluetooth confirmation text
This is a manual merge of the translated strings from the master branch
of the 'bluetooth_pairing_shares_phonebook' string. The fix for this bug
requires these translations because the fix was to slightly reword the
confirmation text in the pairing dialog and stop substituting in the
device name.

Bug: 62672248
Test: manual
Change-Id: I829baada46cdb2314021d8772feb8811072e41d8
(cherry picked from commit 732b65ca4de0bf9cedeb3f2369a2ca596e78cfd3)
Signed-off-by: Albert I <kras@raphielgang.org>